+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.2.0
+
+* 📦 Viewerlist in Explorer
+* 📦 Viewer counter updates every 3m
+* 📦 New Command to clear the "internal" Config file which stores oauth and client-id
+* 👌 Using new Twitch API
+* 🐛 Fixed some bugs
+
+### Known Issue
+
+* When ClientID is not provided you get an error
+
 ## [1.1.0]
 
 * 🌟 Add an Icon to send a Message to twitch chat